+++ app/src/main/AndroidManifest.xml
@@ -1,10 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    package="com.example.financialmanagement">
+    xmlns:tools="http://schemas.android.com/tools">
+
     <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED" />
+
     <application
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
@@ -15,23 +16,41 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.FinancialManagement"
         tools:targetApi="31">
-
+        <activity
+            android:name=".activity.ReportActivity"
+            android:exported="false" />
+        <activity
+            android:name=".activity.SubCategoryActivity"
+            android:exported="false" />
+        <activity
+            android:name=".activity.WalletActivity"
+            android:exported="false" />
+        <activity
+            android:name=".activity.ChangePasswordActivity"
+            android:exported="false" />
         <activity
             android:name=".activity.LoginActivity"
             android:exported="false" />
         <activity
             android:name=".activity.RegisterActivity"
             android:exported="false" />
-
         <activity
             android:name=".activity.AccountActivity"
             android:exported="false" />
         <activity
             android:name=".activity.BudgetActivity"
             android:exported="false" />
-        
         <activity
-            android:name=".MainActivity"
+            android:name=".activity.AddTransactionActivity"
+            android:exported="false" />
+        <activity
+            android:name=".activity.TransactionActivity"
+            android:exported="false" />
+        <activity
+            android:name=".activity.OverviewActivity"
+            android:exported="false" />
+        <activity
+            android:name=".activity.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -40,7 +59,13 @@
             </intent-filter>
         </activity>
 
-
+        <receiver android:name=".notification.NotificationReceiver" />
+        <receiver android:name=".notification.BootReceiver"
+            android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.BOOT_COMPLETED" />
+            </intent-filter>
+        </receiver>
     </application>
 
 </manifest>
