+++ app/src/main/java/com/example/financialmanagement/activity/ChangePasswordActivity.java
@@ -10,7 +10,7 @@ import android.widget.EditText;
 import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
-import com.example.financialmanagement.MainActivity;
+
 import com.example.financialmanagement.R;
 import com.example.financialmanagement.db.DBHelper;
